% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AfterSl1p-package.R
\docType{package}
\name{AfterSl1p-package}
\alias{AfterSl1p}
\alias{AfterSl1p-package}
\title{AfterSl1p: Generate Summary Graphics and Basic Analysis of 16s Data}
\description{
A package of convenience functions to visualize microbiome data. It uses phyloseq and tidyverse/ggplot2 tools. Mostly just a collection of functions I have found useful over the years.
}
\author{
\strong{Maintainer}: JC Szamosi \email{szamosjc@mcmaster.ca}

Authors:
\itemize{
  \item Shahrokh Shekarriz \email{shekas3@mcmaster.ca}
}

}
\keyword{internal}
