% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_ord.R
\name{axis_num}
\alias{axis_num}
\title{Get the axis number
Take a character vector of the form 'Axis.N' where N is a number and return
a numeric vector of N in the same order}
\usage{
axis_num(AxisX)
}
\arguments{
\item{AxisX}{The character vector of axes. Must only contain values of the
form 'Axis.N'}
}
\description{
Get the axis number
Take a character vector of the form 'Axis.N' where N is a number and return
a numeric vector of N in the same order
}
