% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{df_glom}
\alias{df_glom}
\title{Like tax_glom, but for data frames}
\usage{
df_glom(phy_df, IDcol = "X.SampleID", rank = "Phylum",
  abunds = "Abundance", dada2 = FALSE)
}
\arguments{
\item{phy_df}{A phyloseq data frame, as generated by \code{psmelt()} or
\code{make_phy_df()}}

\item{IDcol}{The column name of the sample IDs}

\item{rank}{The taxonomic rank to glom at}

\item{abunds}{The column name of the abundances to sum}

\item{dada2}{(\code{FALSE}) Specifies whether the data came from dada2 or
sl1p. Determines what the taxonomic ranks are.}

\item{tots}{The desired column name of the summed (glommed) abundances}
}
\description{
\code{df_glom} take totals within sample at a given taxonomic rank.
}
