% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{lddf_check}
\alias{lddf_check}
\title{Check the inputs of \code{long_distance_df()}}
\usage{
lddf_check(dmat, metadat, idcol = "X.SampleID", diag = FALSE,
  suff = c("1", "2"))
}
\description{
For internal use only
}
