% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_distance_df.R
\name{lddf_work}
\alias{lddf_work}
\title{Does the actual gathering and spreading without testing assumptions}
\usage{
lddf_work(
  dmat,
  metadat,
  idcol = "X.SampleID",
  suff = c("1", "2"),
  distcol = "Distance"
)
}
\arguments{
\item{dmat}{A distance matrix or other diagonal matrix object with sample
names as row and column names.}

\item{metadat}{A data frame or data frame-like object with the data set's
metadata}

\item{idcol}{(\code{'X.SampleID'}.) A string. The column in \code{metadat}
that holds the sample names. Sample names should match the row/column namse
of the distance matrix. If there are samples in the metadata data frame
that are missing from the distance matrix, they will be excluded with a
warning. If there are samples in the distance matrix that are missing from
the metadata, you will get an error.}

\item{suff}{(\code{c('1','2')}.) A character vector of length 2. The suffixes
to be appended to the metadata column names in the output. The two elements
must not be identical.}

\item{distcol}{(\code{'Distance'}.) A string. The desired column name for the
distance column in your long data frame. Only here to avoid clashes with
existing metadata column names.}
}
\description{
\code{lddf_work} Does the actual gathering, spreading, and joining associated
with making the lddf, but without checking if the distance matrix is sensible
or removing diagonals and repeats. This is for when you know what you're
doing and have trimmed your distance matrix down to only what you know you
need. Good for permutation tests.
}
