% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{make_pcoa_df}
\alias{make_pcoa_df}
\title{Make a data frame of the ordination if the ordination method is PCoA. Not
exported}
\usage{
make_pcoa_df(ord, physeq, axes)
}
\arguments{
\item{ord}{The ordination object generated by \code{phyloseq::ordinate()}.}

\item{physeq}{The phyloseq object we're making things from.}

\item{axes}{A numeric vector of the axes to include.}
}
\description{
Take a phyloseq objet and generate a data frame of the distance-based
ordination of the samples for plotting when the ordination method was PCoA.
Used internally by \code{make_ord_df()}
}
