% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_graphics.R
\name{make_phy_df}
\alias{make_phy_df}
\title{Generate a Data Frame for Taxon Bar Charts}
\usage{
make_phy_df(physeq, rank = "Genus", cutoff = 0.001)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{rank}{The rank at which to glom taxa. Must be one of 'Genus', 'Family',
'Order', 'Class', 'Phylum'. TODO: make it possible to do this OTU-wise.}

\item{cutoff}{The abundance cutoff below which taxa are grouped into 'Other'.
If you don't want anything grouped into 'Other', set this to 0.}
}
\description{
\code{make_phy_df} generates a data frame that is useful for generating taxon
bar charts.
}
\section{Details}{
 This function takes a phyloseq object and generates a data
  frame that is useful for plotting taxon abundance information. By default
  it propagates taxon assignment information down the tree into unassigned
  leves, and aggregates all taxa below 0.1 percent into a single 'Other'
  category. It is most sensible to use this function with a relative
  abundance object, but if you know what you're doing and are careful about
  the cutoff parameter, you can use it on normalized or raw counts.
}

\section{Value}{
 A data frame similar in structure to that generated by
  \code{psmelt}, but with an 'Other' category added and taxon levels ordered
  for use in plotting.
}

