% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_plt_ord.R
\name{make_rda_df}
\alias{make_rda_df}
\title{Make a data frame of the ordination if the ordination method is RDA. Not
exported.}
\usage{
make_rda_df(ord, physeq, axes)
}
\arguments{
\item{ord}{The ordination object generated by
\code{\link[phyloseq]{ordinate}}.}

\item{physeq}{The phyloseq object we're making things from}

\item{axes}{A numeric vector of the axes to plot}
}
\description{
Make a data frame of the ordination if the ordination method is RDA. Not
exported.
}
\section{}{
NA
}

