% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tax_bar.R
\name{order_levs}
\alias{order_levs}
\title{Order the levels of one factor by the values of another}
\usage{
order_levs(f1, ...)
}
\arguments{
\item{f1}{The factor to re-order. Levels of this factor must be nested within
\code{f2}}

\item{f2}{The factor to use when re-ordering \code{f1}.}
}
\description{
\code{order_levs()} takes two factors, the first of which has values that are
nested within the values of the second, and orders the levels of the first
factor such that they are clustered within the second factor.
}
\section{Value}{
 A factor that has the same levels as f1, but that has been
  re-ordered.
}

