% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_graphics.R
\name{order_taxa}
\alias{order_taxa}
\title{Order Taxon Name Factors}
\usage{
order_taxa(phy_df, rank, abund = "Abundance")
}
\arguments{
\item{phy_df}{A data frame of a phyloseq object, as produced by
\code{psmelt()} or \code{make_phy_df()}.}

\item{rank}{The name of the column to be re-ordered}

\item{abund}{The name of the abundances column. Defaults to 'Abundance'}
}
\description{
\code{order_taxa} reorders the taxon names in a taxon column (e.g. 'Class' or
'Phylum') by the taxon's mean abundance (but always makes sure to put Other
first).
}
\section{Value}{
 A data frame that is identical to the one given, but with the
  specified column re-ordered by its mean abundance
}

