% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_graphics.R
\name{plot_tax_bar}
\alias{plot_tax_bar}
\title{Create a Taxa Bar Chart}
\usage{
plot_tax_bar(taxa_df, rank, colours = NULL, sample = "X.SampleID",
  abund = "Abundance", yscale = "lin")
}
\arguments{
\item{rank}{The taxonomic rank at which to view the data. Must be one of
'Genus', 'Family', 'Order', 'Class', 'Phylum'. May not be a lower rank than
the rank given to \code{make_phy_df()}.}

\item{colours}{A character vector of colour names or hex values. Must have
enough colours to accommodate all your taxa at the appropriate rank. If you
don't provide one, there is an internal one that has 21 colours that the
function will try to use.}

\item{sample}{The name of the sample column in the data frame.}

\item{abund}{The name of the abundance column in the data frame.}

\item{yscale}{Can be either 'lin' or 'sqrt'. The 'sqrt' plot can look weird.}

\item{tax_df}{The data frame used for plotting. Unless you really know what
you're doing, use the data frame output by \code{make_phy_df()}.}
}
\description{
\code{plot_tax_bar} creates a taxa bar chart from the data frame generated by
\code{make_phy_df()}.
}
\section{Details}{

  This function generates a ggplot object that is a
  first-pass, reasonable attempt at a taxon bar chart. The taxa are ordered
  by mean abundance and the 'Other' category is at the top. Because it's a
  regular ggplot object, you can add more geoms or facet it or manipulate it
  however you like using ggplot functions.
}

\section{Value}{

  A ggplot object.
}

