% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_graphics.R
\name{plot_taxon_violin}
\alias{plot_taxon_violin}
\title{Make violin plots of taxon abundance}
\usage{
plot_taxon_violin(phy_df, f1 = NULL, f2 = NULL, col_by = NULL,
  colours = NULL)
}
\arguments{
\item{phy_df}{A data frame constructed from a phyloseq object, as with
\code{psmelt()} (or \code{make_phy_df()})}

\item{f1, f2}{Two columns to facet by. If both are used, the first will be
rows and the second will be columns.}

\item{col_by}{A column to colour points by. If this column has more than 21
unique values, you must provide your own list of colours with the
\code{colours} parameter.}

\item{colours}{An optional vector of colours to use when colouring the
points.}
}
\description{
Make violin plots of taxon abundance
}
