% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{plt_ord}
\alias{plt_ord}
\title{Plot the ordination}
\usage{
plt_ord(ord_long, colour = NULL, shape = NULL, pt_alph = 0.7)
}
\arguments{
\item{ord_long}{The ordination data frame produced by \code{make_ord_df}}

\item{colour}{The name of the column to use to colour the points.}

\item{shape}{The name of the coloumn governing the shape of the points.}

\item{pt_alph}{The transparency of the points. Default is 0.7}
}
\description{
Make a multi-axis PCoA plot from the data frame
}
