% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_depth.R
\name{prd_check}
\alias{prd_check}
\title{Check inputs for plot_read_depth()}
\usage{
prd_check(sample_sum_df, cvar, xvar)
}
\arguments{
\item{sample_sum_df}{data frame of sample sums and sample data}

\item{cvar, xvar}{passed through from \code{plot_read_depth()}}
}
\description{
Checks the inputs of \code{plot_read_depth()}
}
