% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_depth.R
\name{prd_clrs}
\alias{prd_clrs}
\title{Deal w colour vector for plot_read_depth()}
\usage{
prd_clrs(sample_sum_df, cvar, clrs)
}
\arguments{
\item{cvar, clrs}{passed through from \code{plot_read_depth()}}
}
\description{
Deal w colour vector for plot_read_depth()
}
