% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_graphics.R
\name{prop_tax_down}
\alias{prop_tax_down}
\title{Propagate taxon information down an entire \code{tax_table} in a phylose
object}
\usage{
prop_tax_down(physeq, dada2)
}
\arguments{
\item{physeq}{a phyloseq object with a filled \code{tax_table} slot.}

\item{dada2}{a flag to indicate whether the data tables were generated in
dada2 or not. If not, it's assumed they were generated in QIIME 1. Just
gets passed to \code{prop_tax_tab()}.}
}
\description{
\code{prop_tax_down} Takes the taxon assignment from the lowest assigned
level in an OTU's assignment and fills it in to all the lower, unresolved
fields.
}
\section{Details}{
 Often taxon cannot be assigned all the way down to genus
  from the limited information available from a given variable region of 16S
  (or other marker gene). If an OTU has been assigned to Order
  Enterobacteriales, but the Family and genus cannot be resolved, then the
  Family and Genus fields are left blank by QIIME. This function fills those
  fields with the string 'o_Enterobacteriales', so they are not parsed as NAs
  by ggplot and other functions.
}

\section{Value}{
 A phyloseq object whose \code{tax_table} has had NAs replaced
  with the higher level taxon assignments.
}

