% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{rank_abund}
\alias{rank_abund}
\title{Rank Taxa by Abundance}
\usage{
rank_abund(phy_df, vars = NULL, bases = NULL, abund = "Abundance",
  ranks = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus",
  "Species"))
}
\arguments{
\item{phy_df}{A dataframe of a phyloseq object, like that generated by
\code{psmelt()} or \code{make_phy_df()}}

\item{vars}{(\code{NULL}) A character vector of grouping variables from which
the baseline values are chosen to define the abundance ordering. If it is
\code{NULL}, the ordering will be based on mean abundances in the whole
data frame.}

\item{bases}{(\code{NULL}) A character vector of baseline values for the
variables given in \code{vars}. The ordering of the taxa will be given
based only on the samples with these baseline values for these variables.
Must be in the same order as vars.}

\item{abund}{(\code{'Abundance'}) The name of the abundance column.}

\item{ranks}{(\code{c('Kingdom','Phylum','Class','Order','Family','Genus','Species')}) A
character vector of taxomonic ranks used by your data set.}
}
\description{
\code{rank_abund()} generates a data frame that's ready to be used by
\code{plot_rank_abund()}
}
\section{Value}{
 A data frame whose taxa have been ranked by their mean
  abundance in the baseline level of some grouping variable or variables.
}

