% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tax_bar.R
\name{rotate_ticks}
\alias{rotate_ticks}
\title{Rotate the x tick labels 90 degrees and position them correctly.}
\usage{
rotate_ticks()
}
\description{
Rotate the x tick labels 90 degrees and position them correctly.
}
