% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_plt_ord.R
\name{make_ord_df}
\alias{make_ord_df}
\title{Make a data frame of the ordination}
\usage{
make_ord_df(
  physeq,
  dist_meth = "bray",
  ord_meth = "PCoA",
  scree_only = FALSE,
  axes = 1:4
)
}
\arguments{
\item{physeq}{A phyloseq object with an OTU table and sample data. The table
should be normalized (rarefied, relative abundance, some kind of log-ratio
transformation) so that the distance metrics will be meaningful. Make sure
that the transformation you have used is appropriate for the distance
method you choose.}

\item{dist_meth}{The distance method to be use. Must be one of the methods
accepted by the \code{phyloseq::distance()} function. Default is 'bray'. If
'jaccard' is used, adds the \code{binary = TRUE} argument. If you have
taken a clr transform of your data and want Aitchison distances, choose
\code{'euclidean'}. Currently, tree-based methods like \code{'unifrac'} and
\code{'wunifrac'} are not supported.}

\item{ord_meth}{The ordination method. Must be one of the methods accepted by
the \code{phyloseq::ordinate()} function. Default is 'PCoA'. Make sure the
ordination method you choose is appropriate for your distance method.}

\item{scree_only}{If \code{TRUE}, this function will print the scree plot of
the requested ordination and then exit. Good for deciding how many axes you
care about. Default is \code{FALSE}.}

\item{axes}{A vector of integers indicating which ordination axes to include
in the data frame. Defaults to \code{1:4}.}
}
\description{
Take a phyloseq object and generate a data frame of the distance-based
ordination of the samples for plotting. This function has been tested with
ordination methods 'PCoA' and 'RDA'. I can't make any promises if you use a
different ordination method.
}
