% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{make_phy_df}
\alias{make_phy_df}
\title{Generate a Data Frame for Taxon Bar Charts}
\usage{
make_phy_df(physeq, rank = "Genus", cutoff = 0.001, indic = FALSE,
  prop = TRUE)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{rank}{The rank at which to glom taxa. Must be one of 'Genus', 'Family',
'Order', 'Class', 'Phylum'. Default is 'Genus'. TODO: make it possible to
do this OTU-wise.}

\item{cutoff}{The abundance cutoff below which taxa are grouped into 'Other'.
If you don't want anything grouped into 'Other', set this to 0. Default is
0.001.}

\item{indic}{a flag to indicate if the taxon names have level indicators.
If FALSE, they are added.}

\item{prop}{Specifies whether taxa need to be propogated down the taxonomy
table (default, TRUE) or if this has already been done.}
}
\description{
\code{make_phy_df} generates a data frame that is useful for generating taxon
bar charts.
}
\section{Details}{
 This function takes a phyloseq object and generates a data
  frame that is useful for plotting taxon abundance information. By default
  it propagates taxon assignment information down the tree into unassigned
  leves, and aggregates all taxa below 0.1 percent into a single 'Other'
  category. This function expects the phyloseq object to be relative
  abundance, and weird things will happen if it is not.
}

\section{Value}{
 A data frame similar in structure to that generated by
  \code{psmelt}, but with an 'Other' category added and taxon levels ordered
  for use in plotting.
}

