% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_ord.R
\name{make_rda_df}
\alias{make_rda_df}
\title{Make a data frame of the ordination if the ordination method is RDA. Not
exported.}
\usage{
make_rda_df(ord, physeq, axes)
}
\arguments{
\item{ord}{The ordination object generated by \code{phyloseq::ordinate()}.}

\item{physeq}{The phyloseq object we're making things from}

\item{axes}{A numeric vector of the axes to plot}
}
\description{
Take a phyloseq object and generate a data frame of the distance-based
ordination of the samples for plotting when the ordination method was
RDA. Used internally by \code{make_ord_df()}
}
