% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{plot_tax_bar}
\alias{plot_tax_bar}
\title{Create a Taxa Bar Chart}
\usage{
plot_tax_bar(taxa_df, rank, colours = NULL, sample = "X.SampleID",
  abund = "Abundance", legloc = "right", yscale = "lin", means = FALSE)
}
\arguments{
\item{rank}{The taxonomic rank at which to view the data. Must be one of
'Genus', 'Family', 'Order', 'Class', 'Phylum'. May not be a lower rank than
the rank given to \code{make_phy_df()}.}

\item{colours}{A character vector of colour names or hex values. Must have
enough colours to accommodate all your taxa at the appropriate rank. If you
don't provide one, there are a few internal vectors that have 21, 31, 60,
or 70 colours that the function will try to use.}

\item{sample}{The name of the sample column in the data frame.}

\item{abund}{The name of the abundance column in the data frame.}

\item{legloc}{Location of the legend. Can be 'right' (default), 'bottom', or
'none' (absent)}

\item{yscale}{Can be either 'lin' or 'sqrt'. The 'sqrt' plot can look weird.}

\item{means}{If \code{TRUE}, sets \code{position = fill} in the
\code{geom_bar()} to constrain the abundances to sum to 1. Good to use if
your \code{sample = } parameter is not actually sample names, but rather
larger categories, to produce means.}

\item{tax_df}{The data frame used for plotting. Unless you really know what
you're doing, use the data frame output by \code{make_phy_df()}.}
}
\description{
\code{plot_tax_bar()} creates a taxa bar chart from the data frame generated
by \code{make_phy_df()}.
}
\section{Details}{
 This function generates a ggplot object that is a
  first-pass, reasonable attempt at a taxon bar chart. The taxa are ordered
  by mean abundance and the 'Other' category is at the top. Because it's a
  regular ggplot object, you can add more geoms or facet it or manipulate it
  however you like using ggplot functions.
}

\section{Value}{
 A ggplot object.
}

