% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{prop_tax_down}
\alias{prop_tax_down}
\title{Propagate taxon information down an entire \code{tax_table} in a phylose
object}
\usage{
prop_tax_down(physeq, indic, dbig = TRUE)
}
\arguments{
\item{physeq}{a phyloseq object with a filled \code{tax_table} slot.}

\item{indic}{a flag to indicate if the taxon names have level indicators. If
FALSE, they are added. Just gets passed to \code{prop_tax_tab()}.}

\item{dbig}{a flag to indicate whether genus names that exist in multiple
families should be disambiguated by appending the family name.}
}
\description{
\code{prop_tax_down} Takes the taxon assignment from the lowest assigned
level in an OTU's assignment and fills it in to all the lower, unresolved
fields.
}
\section{Details}{
 Often taxon cannot be assigned all the way down to genus
  from the limited information available from a given variable region of 16S
  (or other marker gene). If an OTU has been assigned to Order
  Enterobacteriales, but the Family and genus cannot be resolved, then the
  Family and Genus fields are left blank by QIIME. This function fills those
  fields with the string 'o_Enterobacteriales', so they are not parsed as NAs
  by ggplot and other functions.
}

\section{Value}{
 A phyloseq object whose \code{tax_table} has had NAs replaced
  with the higher level taxon assignments.
}

