% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{prop_tax_row}
\alias{prop_tax_row}
\title{Propagate taxon information in a single row of a \code{tax_table} object.}
\usage{
prop_tax_row(taxrow, dada2)
}
\arguments{
\item{taxrow}{A single row in a taxon table}

\item{dada2}{a flag to indicate if the data tables were generated by dada2.
If FALSE, it's assumed they were generated by QIIME 1.}
}
\description{
\code{prop_tax_row} Takes the taxon assignment from the lowest assigned level
in an OTU's assignment and fills it in to all the lower, unresolved fields.
}
\section{Details}{
 Often taxon cannot be assigned all the way down to genus
  from the limited information available from a given variable region of 16S
  (or other marker gene). If an OTU has been assigned to Order
  Enterobacteriales, but the Family and genus cannot be resolved, then the
  Family and Genus fields are left blank by QIIME. This function fills those
  fields with the string 'o_Enterobacteriales', so they are not parsed as NAs
  by ggplot and other functions.
}

\section{Value}{
 A single row \code{tax_table} object to with NAs replaced
  with higher-level taxon assignments
}

\keyword{internal}
