% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_plot_rank_ab.R
\name{rank_abund}
\alias{rank_abund}
\title{Rank Taxa by Abundance}
\usage{
rank_abund(
  phy_df,
  gvars = NULL,
  bases = NULL,
  abunds = "Abundance",
  rank = "Genus",
  IDcol
)
}
\arguments{
\item{phy_df}{A dataframe of a phyloseq object, like that generated by
\code{\link[phyloseq:psmelt]{phyloseq::psmelt()}} or \code{\link[=make_phy_df]{make_phy_df()}}}

\item{gvars}{(\code{NULL}) A character vector of grouping variables from which the
baseline values are chosen to define the abundance ordering. If it is
\code{NULL}, the ordering will be based on mean abundances in the whole data
frame.}

\item{bases}{(\code{NULL}) A character vector of baseline values for the variables
given in \code{gvars}. The ordering of the taxa will be given based only on the
samples with these baseline values for these variables. Must be in the same
order as \code{gvars}.}

\item{abunds}{(\code{'Abundance'}) The name of the abundance column.}

\item{rank}{(\code{'Genus'}) The rank to base the ordering on. Must be a column in
\code{phy_df}}

\item{IDcol}{(\code{'X.SampleID'}) The column name of the sample IDs}
}
\description{
\code{rank_abund()} generates a data frame that's ready to be used by
\code{\link[=plot_rank_ab]{plot_rank_ab()}}
}
\section{Value}{
 A data frame whose taxa have been ranked by their mean
abundance in the user-specified baseline level of some grouping variable or
variables.
}

