% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_graphics.R
\name{remain}
\alias{remain}
\title{Calculate the remainder for the 'Other' section of a phyloseq data frame.}
\usage{
remain(x, tot = 1)
}
\arguments{
\item{x}{A numeric vector}

\item{tot}{The desired total. Default is 1.}
}
\description{
Calculate the remainder for the 'Other' section of a phyloseq data frame.
}
\section{Value}{
 a numeric vector
}

