% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tax_bar.R
\name{subset_order}
\alias{subset_order}
\title{Subset and generate taxon ordering}
\usage{
subset_order(
  phy_df,
  varbs = NULL,
  bases = NULL,
  rank = "Genus",
  abunds = "TotalAbunds"
)
}
\arguments{
\item{phy_df}{A phyloseq data frame, as generated by \code{\link{psmelt}}, but
probably generated by \code{\link{df_glom}}.}

\item{varbs}{(\code{NULL}) A character vector of grouping variables from
which the baseline values are chosen to define the abundance ordering. If
it is \code{NULL}, the ordering will be based on mean abundances in the
whole data frame.}

\item{bases}{(\code{NULL}) A character vector of baseline values for the
variables given in \code{vars}. The ordering of the taxa will be given
based only on the samples with these baseline values for these variables.
Must be in the same order as varbs.}

\item{rank}{(\code{'Genus'}) The rank to base the ordering on.}

\item{abund}{(\code{'Abundance'}) The name of the abundance column.}
}
\description{
\code{subset_order} generates a data frame whose taxon column given by
\code{rank} has been ranked according to its mean abundance in the
\code{abunds} column.
}
